Implement local BLASTn from DB
blastPrep.R does no longer submit to blast. This step has been separated and split into two new parts: local and remote blast. The reson for interrupting the pipeline here is that blast takes a lot of memory / time and should be run separately to conserve resources. On the other hand, if the blast if remote, it will consume no resources at all, but waiting and checking progress is needed.

The local blast arm had now been created but needs some more testing